[ticket-94]: detail page added, ui css updated, components splitted
