+++ src/app/dashboard/dashboard-routing.module.ts
@@ -1,8 +1,25 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 import { DashboardComponent } from './dashboard.component';
+import { DetailComponent } from './detail/detail.component';
+import { MainDashboardComponent } from './main-dashboard/main-dashboard.component';
 
-const routes: Routes = [{ path: '', component: DashboardComponent }];
+const routes: Routes = [
+  {
+    path: '',
+    component: DashboardComponent,
+    children: [
+      {
+        path: '',
+        component: MainDashboardComponent,
+      },
+      {
+        path: 'assets/:id',
+        component: DetailComponent,
+      },
+    ],
+  },
+];
 
 @NgModule({
   imports: [RouterModule.forChild(routes)],
