+++ src/app/dashboard/dashboard.component.html
@@ -1,161 +1,24 @@
-<div class="container py-3">
-  <div class="row">
-    <div class="col-md-4">
-      <p-chart
-        type="pie"
-        [data]="totalCpusRamPie"
-        width="200"
-        [options]="pieOptions"
-      ></p-chart>
-    </div>
-    <div class="col-md-4">
-      <p-chart
-        type="pie"
-        [data]="statusPie"
-        width="200"
-        [options]="pieOptions"
-      ></p-chart>
-    </div>
-    <div class="col-md-4">
-      <p-chart
-        type="pie"
-        [data]="readyPie"
-        width="200"
-        [options]="pieOptions"
-      ></p-chart>
-    </div>
+<div class="row">
+  <div class="col-1">
+    <p-sidebar
+      [(visible)]="visibleSidebar1"
+      [modal]="false"
+      [style]="{ width: '10em' }"
+      [transitionOptions]="'300ms'"
+      [baseZIndex]="10000"
+    >
+      <a routerLink=""><i class="pi pi-home"></i>Home</a>
+    </p-sidebar>
+    <button
+      pButton
+      pRipple
+      type="button"
+      icon="pi pi-bars"
+      class="p-button-rounded p-button-text"
+      (click)="visibleSidebar1 = true"
+    ></button>
   </div>
-  <div class="row search-input">
-    <div class="col-md-4">
-      <span class="p-input-icon-left">
-        <i class="pi pi-search"></i>
-        <input
-          type="text"
-          class="p-inputtext-sm"
-          pInputText
-          placeholder="Search"
-          [(ngModel)]="searchText"
-          (keyup)="onSearch()"
-        />
-      </span>
-      <p-multiSelect
-        [options]="statuses"
-        [(ngModel)]="selectedStatus"
-        optionLabel="name"
-        [style]="{ border: 'none' }"
-        [panelStyle]="{ width: '300px' }"
-        dropdownIcon="pi pi-filter"
-        (onChange)="onStatusChange()"
-      ></p-multiSelect>
-    </div>
-  </div>
-  <div class="row target-assets">
-    <div class="col-md-3 target-card" *ngFor="let target of cpTargetAssets">
-      <p-card styleClass="shadow-2" [style]="{ width: '260px' }">
-        <ng-template pTemplate="header">
-          <div class="row">
-            <div class="col-8" *ngIf="!isFetching">
-              {{ target.name | shorten : 9 }}
-            </div>
-            <div class="col-8" *ngIf="isFetching">
-              <i class="pi pi-spin pi-spinner" style="font-size: 1rem"></i>
-            </div>
-            <div class="col-4">
-              <i
-                class="bi bi-pc-display-horizontal big-icon"
-                [ngClass]="getHighlight(target.status)"
-              ></i>
-            </div>
-          </div>
-        </ng-template>
-        <ng-template pTemplate="content">
-          <div class="row">
-            <div class="col-md-1">
-              <i class="bi bi-cpu"></i>
-            </div>
-            <div class="col-md">
-              <span class="card-label"> CPUs </span>
-              <br />
-              <div *ngIf="!isFetching">
-                {{ target.cpu }}
-              </div>
-              <div *ngIf="isFetching">
-                <i class="pi pi-spin pi-spinner" style="font-size: 1rem"></i>
-              </div>
-            </div>
-            <div class="col-md-1">
-              <i class="bi bi-memory"></i>
-            </div>
-            <div class="col-md">
-              <span class="card-label">Ram</span>
-              <br />
-              <div *ngIf="!isFetching">{{ target.ram | fileSize : ' ' }}MB</div>
-              <div *ngIf="isFetching">
-                <i class="pi pi-spin pi-spinner" style="font-size: 1rem"></i>
-              </div>
-            </div>
-          </div>
-          <div class="row">
-            <div class="col-1">
-              <i class="bi bi-geo-alt"></i>
-            </div>
-            <div class="col-md">
-              <span class="card-label">Location</span>
-              <br />
-              <div *ngIf="!isFetching">
-                {{ target.location }}
-              </div>
-              <div *ngIf="isFetching">
-                <i class="pi pi-spin pi-spinner" style="font-size: 1rem"></i>
-              </div>
-            </div>
-            <div class="col-1">
-              <i class="bi bi-cloud"></i>
-            </div>
-            <div class="col-md">
-              <span class="card-label">Ready</span>
-              <br />
-              <div *ngIf="!isFetching">
-                <div *ngIf="target.isStartable">
-                  <i class="bi bi-check2-circle" style="color: green"></i>
-                </div>
-                <div *ngIf="!target.isStartable && target.isStartable !== null">
-                  <i class="bi bi-x-circle" style="color: red"></i>
-                </div>
-                <div *ngIf="target.isStartable === null">-</div>
-              </div>
-              <div *ngIf="isFetching">
-                <i class="pi pi-spin pi-spinner" style="font-size: 1rem"></i>
-              </div>
-            </div>
-          </div>
-          <div class="row">
-            <div class="col-1">
-              <i class="bi bi-clock-history"></i>
-            </div>
-            <div class="col-md col-xs">
-              <span class="card-label">Status</span>
-              <br />
-              <div *ngIf="!isFetching">
-                {{ target.status | status }}
-              </div>
-              <div *ngIf="isFetching">
-                <i class="pi pi-spin pi-spinner" style="font-size: 1rem"></i>
-              </div>
-            </div>
-          </div>
-        </ng-template>
-        <!-- <hr /> -->
-        <ng-template pTemplate="footer">
-          <!-- <p-button label="Save" icon="pi pi-check"></p-button>
-                      <p-button
-                        label="Cancel"
-                        icon="pi pi-times"
-                        styleClass="p-button-secondary"
-                      ></p-button> -->
-        </ng-template>
-      </p-card>
-      <!-- <p-chart type="pie" [data]="data" ></p-chart> -->
-    </div>
+  <div class="col-11">
+    <router-outlet></router-outlet>
   </div>
 </div>
