+++ src/app/dashboard/main-dashboard/main-dashboard.component.css
@@ -1,0 +1,34 @@
+.big-icon {
+  font-size: xxx-large;
+  display: flex;
+  justify-content: flex-end;
+}
+
+.target-card {
+  margin-top: 0.5rem;
+}
+
+.card-label {
+  color: darkgrey;
+}
+
+.status-running {
+  color: green;
+}
+
+.status-stopped {
+  color: orange;
+}
+
+.status-failed {
+  color: red;
+}
+
+.status-unknown {
+  color: darkgrey;
+}
+
+.search-input,
+.target-assets {
+  margin-top: 1rem;
+}
