+++ src/app/models/target-asset.model.ts
@@ -1,4 +1,4 @@
-export interface TargetAsset {
+export class TargetAsset {
   id: number;
   isStartable: boolean;
   location: string;
@@ -11,4 +11,19 @@ export interface TargetAsset {
   ram: number;
   createdAt: string;
   parentId: number;
+
+  constructor() {
+    this.id = 0;
+    this.isStartable = false;
+    this.location = '';
+    this.owner = '';
+    this.createdBy = '';
+    this.name = '';
+    this.status = '';
+    this.tags = [];
+    this.cpu = 0;
+    this.ram = 0;
+    this.createdAt = '';
+    this.parentId = 0;
+  }
 }
